Corona virus tracker app
Followed this project and studied as I went. Not a great way to be introduced to Spring, but I wanted to get a look at Spring Boot, and get some immediate gratification. This was the first project I followed involving Spring/Spring boot.
I can't vouch for the accuracy of the data being generated. It's taken from another git repository.